Move connection probing to CosmosTestStore
Fixes #24698
