+++ test/EFCore.Cosmos.FunctionalTests/TestUtilities/CosmosTestStore.cs
@@ -4,7 +4,10 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Linq.Expressions;
+using System.Net.Http;
+using System.Net.Sockets;
 using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Azure.Cosmos;
@@ -27,6 +30,7 @@ namespace Microsoft.EntityFrameworkCore.TestUtilities
         private bool _initialized;
 
         private static readonly Guid _runId = Guid.NewGuid();
+        private static bool? _connectionAvailable;
 
         public static CosmosTestStore Create(string name, Action<CosmosDbContextOptionsBuilder> extensionConfiguration = null)
             => new(name, shared: false, extensionConfiguration: extensionConfiguration);
@@ -87,9 +91,71 @@ namespace Microsoft.EntityFrameworkCore.TestUtilities
                 Name,
                 _configureCosmos);
 
+        public static async ValueTask<bool> IsConnectionAvailableAsync()
+        {
+            if (_connectionAvailable == null)
+            {
+                _connectionAvailable = await TryConnectAsync();
+            }
+
+            return _connectionAvailable.Value;
+        }
+
+        private static async Task<bool> TryConnectAsync()
+        {
+            CosmosTestStore testStore = null;
+            try
+            {
+                testStore = CreateInitialized("NonExistent");
+
+                return true;
+            }
+            catch (AggregateException aggregate)
+            {
+                if (aggregate.Flatten().InnerExceptions.Any(IsNotConfigured))
+                {
+                    return false;
+                }
+
+                throw;
+            }
+            catch (Exception e)
+            {
+                if (IsNotConfigured(e))
+                {
+                    return false;
+                }
+
+                throw;
+            }
+            finally
+            {
+                if (testStore != null)
+                {
+                    await testStore.DisposeAsync();
+                }
+            }
+        }
+
+        private static bool IsNotConfigured(Exception exception)
+            => exception switch
+            {
+                HttpRequestException re => re.InnerException is SocketException // Exception in Mac/Linux
+                    || (re.InnerException is IOException ioException && ioException.InnerException is SocketException), // Exception in Windows
+                _ => exception.Message.Contains(
+                    "The input authorization token can't serve the request. Please check that the expected payload is built as per the protocol, and check the key being used.",
+                    StringComparison.Ordinal),
+            };
+
         protected override void Initialize(Func<DbContext> createContext, Action<DbContext> seed, Action<DbContext> clean)
         {
             _initialized = true;
+
+            if (_connectionAvailable == false)
+            {
+                return;
+            }
+
             if (_dataFilePath == null)
             {
                 base.Initialize(createContext ?? (() => _storeContext), seed, clean);
@@ -235,6 +301,11 @@ namespace Microsoft.EntityFrameworkCore.TestUtilities
             if (_initialized
                 && _dataFilePath == null)
             {
+                if (_connectionAvailable == false)
+                {
+                    return;
+                }
+
                 if (Shared)
                 {
                     GetTestStoreIndex(ServiceProvider).RemoveShared(GetType().Name + Name);
